requirements/test.txt
-r base.txt
coverage>=5.5, <5.6
django-sslserver==0.22
